move WizardHeader into molecules because it is composed of many atoms
