+++ app/src/organisms/CalibrateDeck/index.tsx
@@ -32,7 +32,7 @@ import {
   INTENT_DECK_CALIBRATION,
 } from '../../organisms/CalibrationPanels'
 import { ModalShell } from '../../molecules/Modal'
-import { WizardHeader } from '../../atoms/WizardHeader'
+import { WizardHeader } from '../../molecules/WizardHeader'
 import { Portal } from '../../App/portal'
 
 import type { StyleProps, Mount } from '@opentrons/components'
