+++ app/src/organisms/CalibratePipetteOffset/index.tsx
@@ -33,7 +33,7 @@ import {
   MeasureTip,
 } from '../../organisms/CalibrationPanels'
 import { ModalShell } from '../../molecules/Modal'
-import { WizardHeader } from '../../atoms/WizardHeader'
+import { WizardHeader } from '../../molecules/WizardHeader'
 import { Portal } from '../../App/portal'
 
 import type { StyleProps, Mount } from '@opentrons/components'
