+++ app/src/organisms/ChangePipette/ClearDeckModal/index.tsx
@@ -6,7 +6,7 @@ import {
   SPACING,
   JUSTIFY_FLEX_END,
 } from '@opentrons/components'
-import { WizardHeader } from '../../../atoms/WizardHeader'
+import { WizardHeader } from '../../../molecules/WizardHeader'
 import { StyledText } from '../../../atoms/text'
 import { PrimaryButton } from '../../../atoms/buttons'
